прикрутил авторизацию + удалил страницу авторизации
